refactor de rutas web y api, creacion y modificacion de controladores
