+++ app/Http/Controllers/IngressDetailController.php
@@ -10,6 +10,11 @@ use App\Log;
 
 class IngressDetailController extends Controller
 {
+    function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     /**
      * Display a listing of the resource.
      *
